feat(session-detail): add SessionResponse and QuestionAnswer components for displaying question responses with various answer types
